% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression}
\usage{
linear_regression(
  y,
  x,
  intercept = T,
  y_bars = F,
  ss = F,
  variances = F,
  t_test = F,
  f_test = F,
  hat_matrix = F,
  residuals = F
)
}
\arguments{
\item{y}{a vector, outcomes.}

\item{x}{a numerical matrix, each row represents an observation and each column represents a covariate.}

\item{intercept}{a logical value, TRUE by default, indicating should the function include a intercept for the user.}

\item{y_bars}{a logical value, indicating should the function output the estimated y's.
It is FALSE by default, but will be set TRUE if either
ss, variances, t_test, f_test or residuals
is set TRUE}

\item{ss}{a logical value, indicating should the function output sum of squares.
It is FALSE by default, but will be set TRUE if either
variances, f_test or t_test
is set TRUE}

\item{variances}{a logical value, indicating should the function output variances.
It is FALSE by default, but will be set TRUE if
t_test
is set TRUE}

\item{t_test}{a logical value, indicating should the function output results of t test.}

\item{f_test}{a logical value, indicating should the function output results of f test.}

\item{hat_matrix}{a logical value, indicating should the function output the hat matrix.}

\item{residuals}{a logical value, indicating should the function output residuals.}
}
\value{
the regression result, a list
}
\description{
linear_regression
}
\examples{
linear_regression(c(3,5,6),c(1,2,3))
linear_regression(mtcars_mpg, mtcars_covariates)

}
