% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_covariates}
\alias{mtcars_covariates}
\title{mtcars_covariates}
\format{An object of class \code{matrix} with 32 rows and 10 columns.}
\usage{
mtcars_covariates
}
\description{
The data was extracted from the 1974 Motor Trend US magazine,
and comprises 10 aspects of automobile design.
}
\details{
\describe{
 \item{cyl}{Number of cylinders}
 \item{disp}{Displacement (cu.in.)}
 \item{hp}{Gross horsepower}
 \item{drat}{Rear axle ratio}
 \item{wt}{Weight (1000 lbs)}
 \item{qsec}{1/4 mile time}
 \item{vs}{Engine (0 = V-shaped, 1 = straight)}
 \item{am}{Transmission (0 = automatic, 1 = manual)}
 \item{gear}{Number of forward gears}
 \item{carb}{Number of carburetors}
}
}
\keyword{datasets}
