% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_mpg}
\alias{mtcars_mpg}
\title{mtcars_mpg}
\format{An object of class \code{matrix} with 32 rows and 1 columns.}
\usage{
mtcars_mpg
}
\description{
#'The data was extracted from the 1974 Motor Trend US magazine,
and comprises fuel consumption.
}
\details{
\describe{
 \item{mpg}{Miles/(US) gallon}
}
}
\keyword{datasets}
